extract filter options from the SearchForm monolith
